Moved instanceof trait check to the ExistingClassInInstanceOfRule.
